+++ packages/flutter_meedu/lib/flutter_meedu.dart
@@ -2,3 +2,4 @@ export 'consumer.dart';
 export 'notifiers.dart';
 export 'providers.dart';
 export 'rx.dart';
+export 'screen_utils.dart';